Partial fix for major bug in identifying vtables
Vtable identification erroneously allowed the third entry of a vtable to
be 0, which caused many vtables to be missed. This has been partially
corrected, but we should still verify that all methods in a vtable point
to code. This change also exposed a new issue: many classes are now
identified as having multiple corresponding vtables.

Also introduced in this change is partial functionality to rename
vtable entries based on the symbol names in the superclass's vtable.
Currently this functionality is broken due to pruning of classes with
multiple vtables.